# Fix hash join with outer joins and zero/NULL values