--- a/src/lib/operators/join_hash.cpp
+++ b/src/lib/operators/join_hash.cpp
@@ -239,7 +239,7 @@ class JoinHash::JoinHashImpl : public AbstractJoinOperatorImpl {
 
     // Depiction of the hash join parallelization (radix partitioning can be skipped when radix_bits = 0)
     // ===============================================================================================
-    // We have to data paths, one for left side and one for right input side. We can prepare (i.e.,
+    // We have two data paths, one for left side and one for right input side. We can prepare (i.e.,
     // materialize(), build(), etc.) both sides in parallel until the actual join takes place.
     // All tasks might spawn concurrent tasks themselves. For example, materialize parallelizes over
     // the input chunks and the following steps over the radix clusters.