--- a/src/lib/operators/join_hash.cpp
+++ b/src/lib/operators/join_hash.cpp
@@ -118,7 +118,7 @@ class JoinHash::JoinHashImpl : public AbstractJoinOperatorImpl {
   // Determine correct type for hashing
   using HashedType = typename JoinHashTraits<LeftType, RightType>::HashType;
 
-  size_t _calculate_radix_bits() {
+  size_t _calculate_radix_bits() const {
     /*
       Setting number of bits for radix clustering:
       The number of bits is used to create probe partitions with a size that can