--- a/src/lib/operators/join_hash.hpp
+++ b/src/lib/operators/join_hash.hpp
@@ -1,10 +1,5 @@
 #pragma once
 
-#include <memory>
-#include <string>
-#include <utility>
-#include <vector>
-
 #include "abstract_join_operator.hpp"
 #include "types.hpp"
 #include "utils/assert.hpp"