--- a/src/test/operators/join_hash_types_test.cpp
+++ b/src/test/operators/join_hash_types_test.cpp
@@ -1,7 +1,5 @@
-#include "../base_test.hpp"
 #include "gtest/gtest.h"
 
-#include "operators/join_hash.hpp"
 #include "operators/join_hash/join_hash_steps.hpp"
 
 namespace opossum {