This PR fixes #1214.

Some of the changes:
* Making NULL consideration a template parameter (and that's something I'd like to get feedback on)
* Add a test that performs an outer join on a column with zero and NULL values. This test would fail in the current master.
* We now longer store the hash values. I have some veeery rough number of string joins and it makes a difference here, but it's not enough to keep the hashed value around for now.
* Cleaning up some tests (we now test join operator and the hash step functions separately)

Apart from that I had an issue with Clang and "invalid explicitly-specified argument" (296@join_hash.cpp). If anybody has an idea to solve that, please tell me.

Last but not least: performance does improve (should be mainly the removed hash values). I'll run TPC-H over night to measure the impact.