@Bouncner 2018-11-08
Performance: pretty much no changes except from two queries.

TPC-H Q9 is supposed to be a join-heavy query if I remember correctly. What should have changed in Q22 is not clear to me.