@Bouncner 2018-11-09
So for documentation reasons here the results which changes affected the compile times and performance how much.

Compile times:
```
Before: 27s   src/lib/operators/join_hash.cpp
After:  37s   src/lib/operators/join_hash.cpp
```

Micro benchmarks:
```
Initial >>
BM_Join_SmallAndSmall<JoinHash>_median        123,452 ns
BM_Join_SmallAndBig<JoinHash>_median      322,140,684 ns
BM_Join_MediumAndMedium<JoinHash>_median   92,972,600 ns
Hash no longer stored >>
BM_Join_SmallAndSmall<JoinHash>_median        126,838 ns
BM_Join_SmallAndBig<JoinHash>_median      269,045,337 ns
BM_Join_MediumAndMedium<JoinHash>_median   71,167,068 ns
NULLability as template parameter >>
BM_Join_SmallAndSmall<JoinHash>_median        122,362 ns
BM_Join_SmallAndBig<JoinHash>_median      253,440,072 ns
BM_Join_MediumAndMedium<JoinHash>_median   70,601,544 ns
```

So for small hash joins (looking at you, OLTP) avoiding the NULL handling checks bringt some advantages. For larger joins, no longer storing the hash has a nice impact.